+++ src/main/java/com/app/conjugation/repository/TenseRepository.java
@@ -10,10 +10,10 @@ import com.app.conjugation.model.Tense;
 import com.app.conjugation.model.TenseDTO;
 
 @Repository
-public interface TenseRepository extends JpaRepository<Tense, Long> {
+public interface TenseRepository extends JpaRepository<Tense, Integer> {
 	
 	//	@Query("select t from Tense t where t.language.id = :id")
 	@Query("select new com.app.conjugation.model.TenseDTO(t.id, t.name) from Tense t where t.language.id = :id")
-    List<TenseDTO> findByLanguageId(@Param("id") Long id);
+    List<TenseDTO> findByLanguageId(@Param("id") Integer id);
 
 }
